[Bleeding] Tweak NCP's y parameter for onGround seeking, random formatting for chat checks.
This is a small commit which could  actually change a lot of things.
The previous yOnGround value was too high, causing some moves to be detected as "on ground" earlier than expected.

This will affect all moving checks, including NoFall; workarounds which contain ground conditions may need to get reviewed; lowjump-post will likely be affected, also [To be tested]