I don't know what I'm doing
Just an attempt to fix wrong calculation in the first move after landing. Might wrong, but it give the idea why. Let understand client code different way. moveRelative() -> move() -> friction.
Assume everything after move() is the preparation for the next tick not the current one. So the right order is lastfriction -> moveRelative() -> move().