More changes (+)
* Further clarify docs and comments, cleanup.

* Reoder some event-listeners (especially in the movinglistener).

* Net checks: also request a setback on flyingfrequency violations (should prevent server-sided freeze [...]).

* Do still interpret packets for outgoing positions if flyingfrequency is disabled.

* Fix false positives with ImpossibleHit.

* Remove legacy boolean flag for 1.5 inventory tweaks (fastclick).

* Don't allow players to stand on tilted drip leaves (onGround judgement).

* SurvivalFly: reset bunnyhop delay on slopes and stairs, tweak slowfalling handling (only apply slowfall gravity if the player is actually falling [not coming from ground etc...]).

* Remove experimental Improbable feeding with VL frequency hook.

* Rework inventorymove:
Instead of checking during InventoryClickEvent(s), we check players on PlayerMoveEvents and if the inventory is open when it shouldn't be possible, it will be forcibly (and silently) closed, without generating any violation or alert .
This makes the check functionally identical to Open, so it has been merged in it.
Not 100% done. Still needs the speed prediction to function, in order to determine friction speeds.

* Set all config values to build 154 from now on.