FastConsume not so special anymore, do check for client version in some places.
Integrate FastConsume as an ordinary check without checking for specialities/availability
(PlayerItemConsumeEvent was introduced in 1.5 and we do not support versions lower than that anymore, so the availability check has no reason to still stay).

With the client-version hook, we can now finally check for client version in certain places. This isn't really possible for blocks as they are defined globally, and not per-player. Would require a big refactor to support it.

* Revert GodMode formatting and comments.

* Ensure ClientVersion is not set to UKNOWN if protcol-hack plugins are not present on the server, but ProtocolLib is.

* Document stuff